Split core.py into modules and tighten typing (#63)
* Split core.py into core, _file, and _constants modules

core.py had grown to ~1730 lines holding three concerns in one file: the
GoogleDriveFileSystem class, the GoogleDriveFile buffered-file class, and a
block of module constants plus shared helpers.

Extract two modules, keeping the GoogleDriveFile class and moved helpers
byte-for-byte identical (behavior-preserving move):

- _constants.py: module constants, type aliases, MultipleFilesError, and the
  two pure helpers both layers share (_normalize_path, _finfo_from_response).
  Imports nothing from the package except .types, so it never joins a cycle.
- _file.py: GoogleDriveFile plus its private helpers (_should_retry_status,
  _with_supports_all_drives, _parse_range_end). Imports GoogleDriveFileSystem
  only under TYPE_CHECKING.

No re-exports: each symbol now lives in exactly one module, and the tests
import moved symbols from their new homes instead of from .core. The public
API (gdrive_fsspec.GoogleDriveFileSystem) is unchanged.


* Type open() precisely and enforce upload invariants

Follow-up typing pass on the module split. fsspec ships no py.typed, so a
GoogleDriveFile opened via fs.open() was inferred as a text-mode file and
f.write(bytes) tripped bad-argument-type across the integration tests
(previously silenced with 36 per-call pyrefly ignores).

- Override open() with overloads: an uncompressed binary open ("rb"/"wb")
  is typed as a concrete GoogleDriveFile, so read/write are known to take
  and return bytes; text mode and compression fall through to the base
  wrappers and are typed Any. Behavior is unchanged (delegates to super()).
  Removes all 36 bad-argument-type ignores.
- Guard the resumable-upload invariants with explicit raises instead of
  relying on pyrefly-ignored None arithmetic: _upload_chunk now raises if
  offset/location are unset, and _fetch_range raises if file_id is unset
  (rather than sending fileId=None). Raises, not asserts, so they survive
  python -O.
- Initialize _media_object unconditionally in __init__ so it is not a
  branch-conditional attribute.

pyrefly suppressions drop from 44 to 4 (the remainder are irreducible
untyped-fsspec cases). No asserts in src/.


* Type info() as FileInfo and type the capabilities fields

info() always builds a FileInfo but returned dict[str, Any], forcing a
cast() round-trip at its only typed consumer (GoogleDriveFile.__init__) and
leaving the Drive fields it carries untyped.

- Narrow info() to -> FileInfo and drop the two internal dict[str, Any]
  downcasts; drop the now-redundant cast(FileInfo, fs.info(...)) in _file.py.
- Give capabilities and driveId real types (new _Capabilities TypedDict) so
  the shared-drive permission checks in _rm type-check on known value types
  instead of untyped extra-items.

Narrowing the return needs a [bad-override] ignore (fsspec's base info() is
dict[str, Any]). That ignore is scoped to the error code, so on its own it
would also mask a future fsspec change to info()'s return type. A
TYPE_CHECKING pin next to the override asserts the base still returns a
Mapping[str, Any]; if fsspec changes it, that line fails as [bad-assignment]
on its own — a different code the ignore cannot suppress — so the drift
surfaces instead of hiding.

Net: three cast() calls removed, one suppression added, info() honestly
typed. pyrefly 5 suppressed / 0 errors.


* remove comment block

* remove comment block

* Honor end=None as through-EOF in _fetch_range

_fetch_range documents end=None as "fetch through the end", but the range
logic did `end = end or 0` and emitted `bytes=<start>--1` — a malformed
range — whenever start was set and end was None. fsspec's read path never
triggers this (its caches always pass a concrete integer end), so it was
latent, but the code contradicted its own docstring and was a trap for any
direct caller.

Emit an open-ended `bytes=<start>-` when end is None, and only compute the
inclusive end (`end - 1`) when end is provided. Adds a test for the
open-ended case.


---------
